+++ src/modules/components/button-scroll-up/style.tsx
@@ -1,7 +1,7 @@
 import { styled } from "@mui/material/styles";
 import { SECONDARY_COLOR } from "../../../styles/colors";
 import { IconButton } from "@mui/material";
-import { SECONDARY_COLOR_DARK } from "./../../../styles/colors";
+import { SECONDARY_COLOR_DARK } from "../../../styles/colors";
 
 export const Root = styled(({ show, children, ...rest }: any) => (
   <IconButton {...rest}>{children}</IconButton>
@@ -15,6 +15,7 @@ export const Root = styled(({ show, children, ...rest }: any) => (
   height: 56px;
   transition: transform 0.1s ease-in-out;
   transform: ${({ show }) => (show ? "translateX(0)" : "translateX(100px)")};
+  z-index: 200;
 
   &:hover {
     background-color: ${SECONDARY_COLOR_DARK};