+++ src/modules/components/color-picker/index.tsx
@@ -1,29 +1,51 @@
 import { useCallback, useEffect, useState } from "react";
-import SketchPicker from "react-color/lib/components/sketch/Sketch";
+// import { ColorPicker as ReactColorPicker, useColor } from "react-color-palette";
 import strings from "../../../localization";
-import { ColorPickerSpan, Container, Group, SubmitButton } from "./styles";
+import {
+  ColorPickerStyleOverride,
+  PalleteItem,
+  RGBContainer,
+  RGBContainerLabel,
+} from "./styles";
+import { HexColorPicker } from "react-colorful";
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  Grid,
+  InputAdornment,
+  TextField,
+  useMediaQuery,
+  useTheme,
+} from "@mui/material";
+import { DEFAULT_PALLETE, hexToRgb, rgbToHex } from "./utils";
 
 interface IColorPickerProps {
   initialColor: any;
-  onChangeComplete: any;
   onCancel: () => void;
+  onChooseColor: (selectedColor: string) => void;
   id?: string;
 }
 
 const ColorPicker = ({
-  initialColor,
-  onChangeComplete,
+  initialColor = "#ffffff",
   onCancel,
+  onChooseColor,
   id,
 }: IColorPickerProps) => {
-  const [currentColor, setCurrentColor] = useState();
+  const theme = useTheme();
+  const isSmallerThanSM = useMediaQuery(theme.breakpoints.down("sm"));
+
+  const [color, setColor] = useState(initialColor);
+  const { r, g, b } = hexToRgb(initialColor);
+  const [colorRGB, setColorRGB] = useState({ r, g, b });
 
   const escFunction = useCallback((event: any) => {
     if (event.keyCode === 27) {
       onCancel();
     }
-    setCurrentColor(initialColor);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {
@@ -34,41 +56,122 @@ const ColorPicker = ({
     };
   }, [escFunction]);
 
-  const handleChangeComplete = (color: any) => {
-    setCurrentColor(color);
-  };
+  useEffect(() => {
+    setColorRGB(hexToRgb(color));
+  }, [color]);
 
-  const onSubmitColor = (color: any) => {
-    onChangeComplete(color);
-    setCurrentColor(color);
+  const handleOnChange = (selectedColor: string) => {
+    setColor(selectedColor);
   };
 
   return (
-    <ColorPickerSpan
-      onClick={(event: any) => {
-        event.preventDefault();
-      }}
+    <Dialog
       id={id}
+      open={true}
+      fullWidth
+      fullScreen={isSmallerThanSM}
+      maxWidth="xs"
     >
-      <Container>
-        <SketchPicker
-          color={currentColor}
-          onChangeComplete={(color: any) => {
-            handleChangeComplete(color);
-          }}
-        />
-        <Group>
-          <SubmitButton onClick={onCancel}>{strings.cancel}</SubmitButton>
-          <SubmitButton
-            onClick={() => {
-              onSubmitColor(currentColor);
+      <DialogTitle>{strings.colorPicker}</DialogTitle>
+      <DialogContent>
+        <ColorPickerStyleOverride>
+          <HexColorPicker color={color} onChange={handleOnChange} />
+        </ColorPickerStyleOverride>
+        <Grid container mt="24px" gap="8px">
+          {DEFAULT_PALLETE.map((color: string) => (
+            <PalleteItem
+              color={color}
+              key={color}
+              onClick={() => {
+                setColor(color);
+              }}
+            />
+          ))}
+        </Grid>
+        <Grid container mt="32px">
+          <TextField
+            label={strings.color + " (HEX)"}
+            placeholder={"#ffffff"}
+            type="text"
+            fullWidth
+            variant="outlined"
+            value={color.replace("#", "")}
+            onChange={({ target: { value } }: any) => {
+              setColor("#" + value);
             }}
-          >
-            {strings.save}
-          </SubmitButton>
-        </Group>
-      </Container>
-    </ColorPickerSpan>
+            InputLabelProps={{ shrink: true }}
+            InputProps={{
+              startAdornment: (
+                <InputAdornment position="start">#</InputAdornment>
+              ),
+            }}
+          />
+        </Grid>
+        <RGBContainer direction="row" gap="16px">
+          <RGBContainerLabel>{strings.color} RGB</RGBContainerLabel>
+          <TextField
+            type="number"
+            fullWidth
+            variant="outlined"
+            value={colorRGB.r}
+            onChange={({ target: { value } }: any) => {
+              const hex = rgbToHex(Number(value), colorRGB.g, colorRGB.b);
+              setColor(hex);
+            }}
+            InputProps={{
+              inputProps: { min: 0, max: 255 },
+              startAdornment: (
+                <InputAdornment position="start">R:</InputAdornment>
+              ),
+            }}
+          />
+          <TextField
+            type="number"
+            fullWidth
+            variant="outlined"
+            value={colorRGB.g}
+            onChange={({ target: { value } }: any) => {
+              const hex = rgbToHex(colorRGB.r, Number(value), colorRGB.b);
+              setColor(hex);
+            }}
+            InputProps={{
+              inputProps: { min: 0, max: 255 },
+              startAdornment: (
+                <InputAdornment position="start">G:</InputAdornment>
+              ),
+            }}
+            InputLabelProps={{ shrink: true }}
+          />
+          <TextField
+            type="number"
+            fullWidth
+            variant="outlined"
+            value={colorRGB.b}
+            onChange={({ target: { value } }: any) => {
+              const hex = rgbToHex(colorRGB.r, colorRGB.g, Number(value));
+              setColor(hex);
+            }}
+            InputProps={{
+              inputProps: { min: 0, max: 255 },
+              startAdornment: (
+                <InputAdornment position="start">B:</InputAdornment>
+              ),
+            }}
+            InputLabelProps={{ shrink: true }}
+          />
+        </RGBContainer>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onCancel}>{strings.cancel}</Button>
+        <Button
+          onClick={() => {
+            onChooseColor(color);
+          }}
+        >
+          {strings.save}
+        </Button>
+      </DialogActions>
+    </Dialog>
   );
 };