+++ src/modules/pages/user-pages/index.tsx
@@ -14,7 +14,7 @@ import Footer from "../../components/footer";
 import { IUser } from "../../../store/user/types";
 import { canCreatePage } from "../../../utils/plan-enablements";
 import { showErrorToast } from "./../../../utils/toast/index";
-import ButtonScrollUp from "../../components/button-scroll-top";
+import ButtonScrollUp from "../../components/button-scroll-up";
 import { UPPER_MEDIUM_GREY } from "../../../styles/colors";
 
 const UserPages = () => {