+++ src/utils/index.ts
@@ -205,3 +205,9 @@ export const isClickableType = (type: ComponentType) =>
   type === ComponentType.Icon ||
   type === ComponentType.Launch ||
   type === ComponentType.ProgressBar;
+
+export const clamp = (value: number, min: number, max: number) => {
+  if (value < min) return min;
+  else if (value > max) return max;
+  else return value;
+};